Add local workspace mode (File System Access API)
Open a folder on the machine running the browser and work on it directly from the frontend, without the Spin backend in the file path. A new Workspace type has a Local variant (alongside Remote) exposing the same list/read/write/create/search operations; the frontend routes every file operation through it based on the project's mode.

local_fs.rs wraps the File System Access API (directory picker, handle traversal, read via Blob, write via the writable stream). web-sys 0.3.105 leaves gaps, so close() is invoked through Reflect and the write promise is awaited before close. idb.rs persists the picked directory handle in IndexedDB and re-requests read/write permission on reload, so local projects survive a page refresh.

The File System Access API is gated behind web-sys's web_sys_unstable_apis cfg, enabled for the wasm32-unknown-unknown target. Local mode needs a real browser, so it is verified by compile/clippy/build rather than headless E2E.

docs: roadmap Phase 5 done, README status (local mode moved to Working).

+++ frontend/src/components/chat_pane.rs
@@ -27,7 +27,7 @@ pub fn ChatPane(
     Effect::new(move || {
         let _ = messages.get();
         if let Some(el) = scroll_ref.get() {
-            el.set_scroll_top(el.scroll_height());
+            el.set_scroll_top(el.scroll_height() as f64);
         }
     });
